third_party/blink/web_tests/fast/multicol/vertical-rl/caret-range-outside-columns-expected.txt

Test that hitting areas outside columns takes us to the right text content inside columns.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

To the right of the first column:
PASS document.caretRangeFromPoint(550,10).startOffset is 0
FAIL document.caretRangeFromPoint(550,70).startOffset should be 2. Was 0.
FAIL document.caretRangeFromPoint(550,109).startOffset should be 3. Was 0.
To the right of the second column:
FAIL document.caretRangeFromPoint(550,110).startOffset should be 12. Was 8.
FAIL document.caretRangeFromPoint(550,170).startOffset should be 14. Was 8.
FAIL document.caretRangeFromPoint(550,209).startOffset should be 15. Was 8.
To the right of the third column:
FAIL document.caretRangeFromPoint(550,210).startOffset should be 24. Was 20.
FAIL document.caretRangeFromPoint(550,270).startOffset should be 26. Was 20.
FAIL document.caretRangeFromPoint(550,309).startOffset should be 27. Was 20.
To the left of the first column:
PASS document.caretRangeFromPoint(0,10).startOffset is 8
FAIL document.caretRangeFromPoint(0,70).startOffset should be 10. Was 8.
FAIL document.caretRangeFromPoint(0,109).startOffset should be 11. Was 8.
To the left of the second column:
PASS document.caretRangeFromPoint(0,110).startOffset is 20
FAIL document.caretRangeFromPoint(0,170).startOffset should be 22. Was 20.
FAIL document.caretRangeFromPoint(0,209).startOffset should be 23. Was 20.
To the left of the third column:
PASS document.caretRangeFromPoint(0,210).startOffset is 32
FAIL document.caretRangeFromPoint(0,270).startOffset should be 34. Was 32.
FAIL document.caretRangeFromPoint(0,309).startOffset should be 35. Was 32.
Above first column:
PASS document.caretRangeFromPoint(550,0).startOffset is 0
PASS document.caretRangeFromPoint(490,0).startOffset is 0
PASS document.caretRangeFromPoint(470,0).startOffset is 4
PASS document.caretRangeFromPoint(450,0).startOffset is 8
Below third column:
PASS document.caretRangeFromPoint(490,350).startOffset is 27
PASS document.caretRangeFromPoint(470,350).startOffset is 31
PASS document.caretRangeFromPoint(450,350).startOffset is 35
In the gap between the first and second columns:
PASS document.caretRangeFromPoint(499,109).startOffset is 3
PASS document.caretRangeFromPoint(479,109).startOffset is 7
PASS document.caretRangeFromPoint(459,109).startOffset is 11
PASS document.caretRangeFromPoint(440,109).startOffset is 11
PASS document.caretRangeFromPoint(499,110).startOffset is 12
PASS document.caretRangeFromPoint(479,110).startOffset is 16
PASS document.caretRangeFromPoint(459,110).startOffset is 20
PASS document.caretRangeFromPoint(440,110).startOffset is 20
In the gap between the second and third columns:
PASS document.caretRangeFromPoint(499,209).startOffset is 15
PASS document.caretRangeFromPoint(479,209).startOffset is 19
PASS document.caretRangeFromPoint(459,209).startOffset is 23
PASS document.caretRangeFromPoint(440,209).startOffset is 23
PASS document.caretRangeFromPoint(499,210).startOffset is 24
PASS document.caretRangeFromPoint(479,210).startOffset is 28
PASS document.caretRangeFromPoint(459,210).startOffset is 32
PASS document.caretRangeFromPoint(440,210).startOffset is 32
PASS successfullyParsed is true

TEST COMPLETE

xxx xxx xxx xxx xxx xxx xxx xxx xxx
